pyproject.toml
[project]
name = "notify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin>=7.1.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "logging>=0.4.9.6",
    "pika>=1.3.2",
    "pybreaker>=1.4.1",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "schema>=0.7.8",
    "uuid>=1.30",
]
